requirements.txt
PyOpenGL
Pygame
numpy
matplotlib